python-backend/requirements.txt
asgiref==3.11.1
Django==5.0.3
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.9.0
psycopg==3.2.10
psycopg-binary==3.2.10
sqlparse==0.5.5
tzdata==2025.3
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0

# Exports Excel/PDF
openpyxl==3.1.2
reportlab==4.0.9
Pillow==10.2.0

# Notifications Firebase
firebase-admin==6.4.0

# QR Codes
qrcode==7.4.2

# Utilitaires
python-dateutil==2.8.2
